split script into separate files
