+++ napi/api.py
@@ -1,0 +1,30 @@
+import os
+from urllib import request
+
+
+def f(z):
+    idx = [0xe, 0x3, 0x6, 0x8, 0x2]
+    mul = [2, 2, 5, 4, 3]
+    add = [0, 0xd, 0x10, 0xb, 0x5]
+
+    b = []
+    for i in range(len(idx)):
+        a = add[i]
+        m = mul[i]
+        i = idx[i]
+
+        t = a + int(z[i], 16)
+        v = int(z[t:t + 2], 16)
+        b.append(("%x" % (v * m))[-1])
+
+    return "".join(b)
+
+
+def build_url(movie_hash):
+    return "http://napiprojekt.pl/unit_napisy/dl.php?l=PL&f={}&t={}&v=other&kolejka=false&nick=&pass=&napios={}".format(
+        movie_hash, f(movie_hash), os.name)
+
+
+def download(movie_hash: str) -> bytes:
+    the_url = build_url(movie_hash)
+    return request.urlopen(the_url).read()
